HTML back-end moved to src/ directory
This move leads to less code repetition as all plugins could utilize
this class.

Also wrapper class WebPage was created to simplify work with web pages.

As the main program does not need the class, the linker may throw it
away - the `-rdynamic' switch prevents that (hope so, Google provided
very little information).